new1/requirements.txt.txt
Flask==2.3.2
SQLAlchemy==2.0.19
python-docx==0.8.11
